requirements.txt
fastembed
flask
flask-cors
groq
faiss-cpu
rank-bm25
python-dotenv
pandas
pymupdf
numpy
gunicorn
